packages/blockchain/pyproject.toml
[project]
name="blockchain"
version = "0.1.0"
description = "Library implementing the blockchain proposed in my thesis"
authors = [{name = "ale18V", email = "alebombarda@icloud.com"}]
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "click>=8.1.8",
    "dependency-injector>=4.44.0",
    "ecdsa>=0.19.0",
    "grpcio>=1.64.3",
    "hypercorn>=0.17.3",
    "loguru>=0.7.3",
    "quart>=0.20.0",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.62.3",
    "mypy>=1.14.0",
    "mypy-protobuf>=3.6.0",
    "protoletariat>=3.3.6",
    "types-protobuf>=5.29.1.20241207",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/blockchain"]


[tool.mypy]
exclude=["generated/", "tests/"]

check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_defs = true 
# disallow_incomplete_defs = true

warn_unused_ignores = true
warn_redundant_casts = true
warn_no_return = true
warn_return_any = true

no_implicit_optional = true
strict_optional = true
strict_equality = true

disallow_subclassing_any = false 
disallow_untyped_decorators = false
allow_redefinition = true 

[[tool.mypy.overrides]]
module = "*.generated.*"
follow_imports = "skip"


